--- src/main/java/kcs/funding/fundingboost/domain/controller/MyPayController.java
@@ -1,32 +1,0 @@
-package kcs.funding.fundingboost.domain.controller;
-
-import kcs.funding.fundingboost.domain.dto.common.CommonSuccessDto;
-import kcs.funding.fundingboost.domain.dto.global.ResponseDto;
-import kcs.funding.fundingboost.domain.dto.request.PaymentDto;
-import kcs.funding.fundingboost.domain.dto.response.MyPayViewDto;
-import kcs.funding.fundingboost.domain.service.MyPayService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/mypay")
-public class MyPayController {
-
-    private final MyPayService myPayService;
-
-    @GetMapping("/order")
-    public ResponseDto<MyPayViewDto> myOrderPayView(@RequestParam(name = "memberId") Long memberId) {
-        return ResponseDto.ok(myPayService.viewOrder(memberId));
-    }
-
-    @GetMapping("/funding")
-    public ResponseDto<MyPayViewDto> myFundingPayView( @RequestParam(name = "memberId") Long memberId){
-        return ResponseDto.ok(myPayService.viewFunding(memberId));
-    }
-
-    @PostMapping("/payment")
-    public ResponseDto<CommonSuccessDto> processMyPayment( @RequestBody PaymentDto paymentDto, @RequestParam("memberId") Long memberId){
-        return ResponseDto.ok(myPayService.pay(paymentDto, memberId));
-    }
-}
